pyproject.toml
[project]
name = "whatsapp-analyzer"
version = "0.1.0"
description = "A tool for analyzing WhatsApp group chat exports"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "pandas>=2.0.0",
    "click>=8.0.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "ruff>=0.11.6",
    "python-fasthtml>=0.12.12",
]
requires-python = ">=3.12"

[project.scripts]
whatsapp-analyzer = "src.cli.main:cli"
whatsapp-web = "src.web.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
